Separate global from local to handle #include issues.
